requirements.txt
dj-database-url==0.5.0
Django==2.0.3
djangorestframework==3.7.7
pkg-resources==0.0.0
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.3
Unipath==1.1
pytest-django
pytest